requirements.txt
certifi
speedtest-cli
pillow
wordcloud
pytz
requests
pyyaml
emoji
email_validator
pypng
pyzbar
pyrogram
redis